Fix #5610 - Nordstrom not loading (#5651)
The history state helper monkey patches the history push state, but our
override throws a DataCloneError when copying the url. This breaks
pushState behaviour. The good news is that we no longer need this hack,
we get URL change KVO events when pushState nav happens. Tested on
youtube.com and a few other sites.